+++ test/LoaderTest.php
@@ -2,7 +2,6 @@
 declare(strict_types=1);
 namespace yii\mustache;
 
-use function PHPUnit\Expect\{expect, it};
 use PHPUnit\Framework\{TestCase};
 use yii\base\{InvalidCallException};
 
@@ -17,28 +16,28 @@ class LoaderTest extends TestCase {
   private $model;
 
   /**
-   * @test Loader::findViewFile
+   * Tests the `Loader::findViewFile
    */
-  public function testFindViewFile(): void {
+  function testFindViewFile(): void {
     $findViewFile = function($name) {
       return $this->findViewFile($name);
     };
 
-    it('should return the path of the corresponding view file', function() use ($findViewFile) {
-      expect($findViewFile->call($this->model, '//view'))->to->equal(str_replace('/', DIRECTORY_SEPARATOR, \Yii::$app->viewPath.'/view.php'));
+    // It should return the path of the corresponding view file.
+      assertThat($findViewFile->call($this->model, '//view'), equalTo(str_replace('/', DIRECTORY_SEPARATOR, \Yii::$app->viewPath.'/view.php'));
     });
 
-    it('should throw an exception if the view file is not found', function() use ($findViewFile) {
-      expect(function() use ($findViewFile) { $findViewFile->call($this->model, '/view'); })->to->throw(InvalidCallException::class);
+    // It should throw an exception if the view file is not found.
+      assertThat(function() use ($findViewFile) { $findViewFile->call($this->model, '/view'); })->to->throw(InvalidCallException::class));
     });
   }
 
   /**
-   * @test Loader::load
+   * Tests the `Loader::load
    */
-  public function testLoad(): void {
-    it('should throw an exception if the view file is not found', function() {
-      expect(function() { $this->model->load('view'); })->to->throw(InvalidCallException::class);
+  function testLoad(): void {
+    // It should throw an exception if the view file is not found.
+      assertThat(function() { $this->model->load('view'); })->to->throw(InvalidCallException::class));
     });
   }