+++ test/LoggerTest.php
@@ -2,7 +2,6 @@
 declare(strict_types=1);
 namespace yii\mustache;
 
-use function PHPUnit\Expect\{expect, it};
 use PHPUnit\Framework\{TestCase};
 use yii\base\{InvalidParamException};
 
@@ -12,11 +11,11 @@ use yii\base\{InvalidParamException};
 class LoggerTest extends TestCase {
 
   /**
-   * @test Logger::log
+   * Tests the `Logger::log
    */
-  public function testLog(): void {
-    it('should throw an exception if the log level is invalid', function() {
-      expect(function() { (new Logger)->log('dummy', 'Hello World!'); })->to->throw(InvalidParamException::class);
+  function testLog(): void {
+    // It should throw an exception if the log level is invalid.
+      assertThat(function() { (new Logger)->log('dummy', 'Hello World!'); })->to->throw(InvalidParamException::class));
     });
   }
 }