+++ test/ViewRendererTest.php
@@ -2,7 +2,6 @@
 declare(strict_types=1);
 namespace yii\mustache;
 
-use function PHPUnit\Expect\{expect, it};
 use PHPUnit\Framework\{TestCase};
 use yii\web\{View};
 
@@ -17,50 +16,50 @@ class ViewRendererTest extends TestCase {
   private $model;
 
   /**
-   * @test ViewRenderer::getHelpers
+   * Tests the `ViewRenderer::getHelpers
    */
-  public function testGetHelpers(): void {
-    it('should return a Mustache helper collection', function() {
-      expect($this->model->helpers)->to->be->instanceOf(\Mustache_HelperCollection::class);
+  function testGetHelpers(): void {
+    // It should return a Mustache helper collection.
+      assertThat($this->model->helpers, isInstanceOf(\Mustache_HelperCollection::class));
     });
   }
 
   /**
-   * @test ViewRenderer::render
+   * Tests the `ViewRenderer::render
    */
-  public function testRender(): void {
+  function testRender(): void {
     $file = __DIR__.'/fixtures/data.mustache';
     $view = new View;
 
-    it('should remove placeholders when there is no corresponding binding', function() use ($file, $view) {
+    // It should remove placeholders when there is no corresponding binding.
       $data = null;
       $output = preg_split('/\r?\n/', $this->model->render($view, $file, $data));
-      expect($output[0])->to->equal('<test></test>');
-      expect($output[1])->to->equal('<test></test>');
-      expect($output[2])->to->equal('<test></test>');
-      expect($output[3])->to->equal('<test>hidden</test>');
+      assertThat($output[0], equalTo('<test></test>');
+      assertThat($output[1], equalTo('<test></test>');
+      assertThat($output[2], equalTo('<test></test>');
+      assertThat($output[3], equalTo('<test>hidden</test>');
     });
 
-    it('should replace placeholders with the proper values when there is a corresponding binding', function() use ($file, $view) {
+    // It should replace placeholders with the proper values when there is a corresponding binding.
       $data = ['label' => '"Mustache"', 'show' => true];
       $output = preg_split('/\r?\n/', $this->model->render($view, $file, $data));
-      expect($output[0])->to->equal('<test>&quot;Mustache&quot;</test>');
-      expect($output[1])->to->equal('<test>"Mustache"</test>');
-      expect($output[2])->to->equal('<test>visible</test>');
-      expect($output[3])->to->equal('<test></test>');
+      assertThat($output[0], equalTo('<test>&quot;Mustache&quot;</test>');
+      assertThat($output[1], equalTo('<test>"Mustache"</test>');
+      assertThat($output[2], equalTo('<test>visible</test>');
+      assertThat($output[3], equalTo('<test></test>');
     });
   }
 
   /**
-   * @test ViewRenderer::setHelpers
+   * Tests the `ViewRenderer::setHelpers
    */
-  public function testSetHelpers(): void {
-    it('should allow arrays as input', function() {
+  function testSetHelpers(): void {
+    // It should allow arrays as input.
       $this->model->helpers = ['var' => 'value'];
 
       $helpers = $this->model->helpers;
-      expect($helpers->has('var'))->to->be->true;
-      expect($helpers->get('var'))->to->equal('value');
+      assertThat($helpers->has('var'), isTrue());
+      assertThat($helpers->get('var'), equalTo('value');
     });
   }